bskycard: separate and reorganize
